[#193] feat : 구성원 페이지 서버렌더링
